updated charge because having decimal * float issues and updated buyer phone number as it invlidated form submit after payment.

+++ seller/views.py
@@ -115,7 +115,7 @@ def charge(request):
 	if form.is_valid():
 		seller_email = form.cleaned_data['email']
 		seller = Seller.objects.get(email=seller_email)
-		price_in_cents = int(seller.unit_price * 100 * float(form.cleaned_data['quantity']))
+		price_in_cents = int(int(seller.unit_price * 100) * float(form.cleaned_data['quantity']))
 		card = request.POST.get('stripeToken')
 		stripe.api_key = settings.STRIPE_KEYS['API_KEY']
 		stripe_account_id = SocialAccount.objects.get(user__id=seller.id, provider='stripe').uid
@@ -139,11 +139,11 @@ def charge(request):
 		sale.save()
 		
 		# send email
-        email_subject = 'Order Confirmation'    
-        email_body = "Dear %s.\n\nYou've got an order from Customer: %s \nAddress: %s\nPhone Number: %s\nQuantity: %.2f\nPlease confirm the order and fulfill it.\n\nThank you." % (seller.first_name, sale.buyer_name, sale.delivery_address, sale.buyer_phone, sale.quantity)
-        send_mail(email_subject, email_body, settings.DEFAULT_FROM_EMAIL, [seller.email], fail_silently=False)
-
-        return render(request, 'order_success.html', {'seller': seller})
+		email_subject = 'Order Confirmation'
+		email_body = "Dear %s.\n\nYou've got an order from Customer: %s \nAddress: %s\nPhone Number: %s\nQuantity: %.2f\nPlease confirm the order and fulfill it.\n\nThank you." % (seller.first_name, sale.buyer_name, sale.delivery_address, sale.buyer_phone, sale.quantity)
+		send_mail(email_subject, email_body, settings.DEFAULT_FROM_EMAIL, [seller.email], fail_silently=False)
+		
+		return render(request, 'order_success.html', {'seller': seller})
 	return render(request, 'order.html', {
         'form': form,
 		'key': settings.STRIPE_KEYS['PUBLIC_KEY']
